completed issues #1 user can create a event.
