+++ server/public/styles.css
@@ -86,3 +86,32 @@ a {
 label {
   font-size: 18px;
 }
+
+.event-card {
+  border: solid white;
+  width: 350px;
+  height: 350px;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
+  transition: 0.4s;
+  margin-top: 20px;
+}
+
+.event-card:hover {
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+}
+
+.allEvents-container {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  justify-content: space-evenly;
+}
+
+h3 {
+  color: #007bff;
+}
+
+.event-text {
+  text-align: center;
+  padding: 70px;
+}
